# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "alpha-visualizer",
    "description": "AlphaForge バックテスト結果の Web 可視化ツール",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/strategies/compare": {
      "get": {
        "summary": "Compare Strategies",
        "operationId": "compare_strategies_api_strategies_compare_get",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "カンマ区切りの strategy_id",
              "title": "Ids"
            },
            "description": "カンマ区切りの strategy_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StrategyComparison"
                  },
                  "title": "Response Compare Strategies Api Strategies Compare Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/strategies/{strategy_id}": {
      "get": {
        "summary": "Get Strategy",
        "operationId": "get_strategy_api_strategies__strategy_id__get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/strategies/{strategy_id}/parameters": {
      "post": {
        "summary": "Save Strategy Parameters",
        "description": "編集済みパラメータを戦略定義へ書き戻す（destructive・UI 側で確認必須）。\n\nvisualizer はファイル・DB を直接書かず、パラメータ差し替え済みの一時\nJSON を ``forge strategy save --force`` に渡して登録を委譲する\n（single-writer 維持。スキーマ検証は forge 側 Pydantic が SSoT で、\nエラーはそのまま表面化させる）。",
        "operationId": "save_strategy_parameters_api_strategies__strategy_id__parameters_post",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveParametersRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveParametersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComparisonEquityCurve": {
        "properties": {
          "dates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dates",
            "default": []
          },
          "values": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Values",
            "default": []
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ComparisonEquityCurve",
        "description": "``StrategyComparison.equity`` の構造（dates / values の配列ペア）。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OptimizationHistoryEntry": {
        "properties": {
          "trial": {
            "type": "integer",
            "title": "Trial"
          },
          "best_sharpe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Sharpe"
          },
          "run_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run At"
          },
          "n_trials": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N Trials"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "trial"
        ],
        "title": "OptimizationHistoryEntry",
        "description": "``StrategyDetail.optimization_history`` の 1 件。"
      },
...
      "SaveParametersRequest": {
        "properties": {
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "maxProperties": 100,
            "minProperties": 1,
            "title": "Parameters"
          }
        },
        "type": "object",
        "required": [
          "parameters"
        ],
        "title": "SaveParametersRequest"
      },
      "SaveParametersResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          },
          "log_tail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Tail"
          }
        },
        "type": "object",
        "required": [
          "status",
          "parameters"
        ],
        "title": "SaveParametersResponse"
      },
      "StrategyComparison": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol",
            "default": ""
          },
          "total_return_pct": {
            "type": "number",
            "title": "Total Return Pct",
            "default": 0.0
          },
          "cagr_pct": {
            "type": "number",
            "title": "Cagr Pct",
            "default": 0.0
          },
          "sharpe_ratio": {
            "type": "number",
            "title": "Sharpe Ratio",
            "default": 0.0
          },
          "sortino_ratio": {
            "type": "number",
            "title": "Sortino Ratio",
            "default": 0.0
          },
          "max_drawdown_pct": {
            "type": "number",
            "title": "Max Drawdown Pct",
            "default": 0.0
          },
          "win_rate_pct": {
            "type": "number",
            "title": "Win Rate Pct",
            "default": 0.0
          },
          "profit_factor": {
            "type": "number",
            "title": "Profit Factor",
            "default": 0.0
          },
          "total_trades": {
            "type": "integer",
            "title": "Total Trades",
            "default": 0
          },
          "is_baseline": {
            "type": "boolean",
            "title": "Is Baseline",
            "default": false
          },
          "equity": {
            "$ref": "#/components/schemas/ComparisonEquityCurve",
            "default": {
              "dates": [],
              "values": []
            }
          },
          "daily_returns": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Daily Returns",
            "default": []
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "StrategyComparison",
        "description": "``GET /api/strategies/compare`` の 1 件。equity / 比較指標を含む。"
      },
      "StrategyDetail": {
        "properties": {
          "strategy_id": {
            "type": "string",
            "title": "Strategy Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters",
            "default": {}
          },
          "indicators": {
            "items": {},
            "type": "array",
            "title": "Indicators",
            "default": []
          },
          "variables": {
            "items": {},
            "type": "array",
            "title": "Variables",
            "default": []
          },
          "entry_conditions": {
            "title": "Entry Conditions"
          },
          "exit_conditions": {
            "title": "Exit Conditions"
          },
          "risk_management": {
            "title": "Risk Management"
          },
          "regime_config": {
            "title": "Regime Config"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/StrategyResultEntry"
            },
            "type": "array",
            "title": "Results",
            "default": []
          },
          "optimization_history": {
            "items": {
              "$ref": "#/components/schemas/OptimizationHistoryEntry"
            },
            "type": "array",
            "title": "Optimization History",
            "default": []
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "strategy_id",
          "name"
        ],
        "title": "StrategyDetail",
        "description": "``GET /api/strategies/{strategy_id}`` 詳細レスポンス。\n\n戦略定義（parameters / indicators / variables / *_conditions /\nrisk_management / regime_config）と、バックテスト履歴 / 最適化履歴を返す。"
      },
      "StrategyResultEntry": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "sharpe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharpe"
          },
          "return_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Pct"
          },
          "max_drawdown_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Drawdown Pct"
          },
          "total_trades": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Trades"
          },
          "run_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run At"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "StrategyResultEntry",
        "description": "``StrategyDetail.results`` の 1 件（バックテスト履歴サマリ）。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}